refactored generation of pka sets
