--- pIChemiSt/test/pka_sets_gauci_expected.json
@@ -1,1 +1,0 @@
-{"acidic": {"D": [4.05, 4.05, 4.05], "C": [9.0, 9.0, 9.0], "E": [4.45, 4.45, 4.45], "Y": [10.0, 10.0, 10.0], "U": [5.43, 5.2, 5.6]}, "basic": {"R": [12.0, 12.0, 12.0], "H": [5.98, 5.98, 5.98], "K": [10.0, 10.0, 10.0]}, "terminus_ionizable": {"A": [7.59, 3.55], "R": [7.5, 3.55], "N": [6.7, 3.55], "D": [7.5, 4.55], "C": [6.5, 3.55], "E": [7.7, 4.75], "Q": [7.5, 3.55], "G": [7.5, 3.55], "H": [7.5, 3.55], "I": [7.5, 3.55], "L": [7.5, 3.55], "K": [7.5, 3.55], "M": [7.0, 3.55], "F": [7.5, 3.55], "P": [8.3599, 3.55], "S": [6.93, 3.55], "T": [6.82, 3.55], "W": [7.5, 3.55], "Y": [7.5, 3.55], "V": [7.44, 3.55], "X": [7.376994999999999, 3.659999999999999], "Z": [7.6, 4.15], "B": [7.1, 4.05], "U": [7.376994999999999, 3.659999999999999]}}
