--- pIChemiSt/test/pka_sets_nozaki_expected.json
@@ -1,1 +1,0 @@
-{"acidic": {"D": [4.0, 4.0, 4.0], "E": [4.4, 4.4, 4.4], "C": [9.5, 9.5, 9.5], "Y": [9.6, 9.6, 9.6]}, "basic": {"K": [10.4, 10.4, 10.4], "R": [12.0, 12.0, 12.0], "H": [6.3, 6.3, 6.3]}, "terminus_ionizable": {"G": [7.5, 3.8], "A": [7.5, 3.8], "S": [7.5, 3.8], "P": [7.5, 3.8], "V": [7.5, 3.8], "T": [7.5, 3.8], "C": [7.5, 3.8], "I": [7.5, 3.8], "L": [7.5, 3.8], "N": [7.5, 3.8], "D": [7.5, 3.8], "Q": [7.5, 3.8], "K": [7.5, 3.8], "E": [7.5, 3.8], "M": [7.5, 3.8], "H": [7.5, 3.8], "F": [7.5, 3.8], "R": [7.5, 3.8], "Y": [7.5, 3.8], "W": [7.5, 3.8], "X": [7.5, 3.8], "Z": [7.5, 3.8], "B": [7.5, 3.8], "U": [7.5, 3.8]}}
